chore(deps): break MTFParser components/helpers cycle
Closes circular dependency #6 (MTFParser.components.ts <-> MTFParserHelpers.ts)
by extracting the shared parseField primitive to a leaf module
MTFParser.types.ts that both halves import.

Pattern: over-decomposed split (Pattern 1 in the cleanup plan). Same
mechanical fix applied in PR #420 (utils/colors, techBaseValidation) and
PR #419 (services repos). Re-export of parseField preserved in
MTFParserHelpers for backward compat with any external callers.

Verification:
- npx madge --circular --extensions ts,tsx src/services/conversion/
  -> 1 cycle remains (BlkExportServiceCore, deferred to PR-B6)
  -> #6 (MTFParser) is gone
- npx tsc --noEmit -> clean
- npx jest src/services/conversion -> 13 suites, 715/715 pass

Unblocks the schema bridge effort (PR-A1) which will edit adjacent
conversion files.
